pyproject.toml
[build-system]
requires = ["hatchling >= 1.26"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/fbi_api"]

[project]
name = "fbi-data-api"
version = "0.3.0"
authors = [
  { name="Ted Jesus C. Chua", email="chuatedjesusc@gmail.com" },
]
description = "Interface for extracting data assets from the FBI Crime Data API."
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "pandas>=3.0.1",
    "python-dotenv>=1.2.2",
    "pyyaml>=6.0.3",
    "requests>=2.32.5",
]
license = "MIT"
license-files = ["LICEN[CS]E*"]

[dependency-groups]
dev = [
    "pytest>=9.0.2",
    "pytest-mock>=3.15.1",
]

[project.urls]
Homepage = "https://github.com/teddythepooh/fbi_api"
Issues = "https://github.com/teddythepooh/fbi_api/issues"
